Data pre-processing + augmentation.
In the first phase of the project, just as expected, data pre-processing was done.  This included very simple process of loading images, resizing and normalizing them.
Later on, I realized that this dataset is highly imbalanced, and therefore we needed to handle this issue. Hence, by manually taking proportional samples of the dataset the issue was refined. Since after the alterations our dataset is not large enough, data augmentation came for help. This process contains horizontal flip and 90-degrees rotation for now, but I'm looking forward to enhancing the data augmentation by introducing some additional tranformations like vertical flip, zoom in, zoom out and etc.
Once our dataset is extended through augmentation techniques, we split our data into training and testing sets, using the ratio of 8:2. In addition, a quarter of training data is dedicated to validation set. (These percentages may be revised based on the future results to improve the model.)